refactor(iota-vm-sdk): split wasm surface into submodules
Break the single src/wasm.rs into a wasm/ module directory:

- mod.rs   — module docs, the wasm-bindgen entry point, and the shared
             b64_decode / err_to_js helpers
- decode.rs — the static, VM-free decode helpers and their result types
- types.rs  — the serde request/result types for simulate
- simulate.rs — the simulate entry point

No behavior change; the JS surface is identical.
